fixed: migrate cadence code first
Closes #1178
